requirements.txt
cryptography==36.0.1
PyJWT==1.7.1
